renamed modify to components
