+++ packages/warriorjs-cli/src/layouts/components/constructSeperator.js
