+++ packages/warriorjs-cli/src/layouts/components/eventMessage.js
