+++ packages/warriorjs-cli/src/layouts/components/floorMap.js
