+++ packages/warriorjs-cli/src/layouts/components/logMessage.js
