+++ packages/warriorjs-cli/src/layouts/components/success.js
