+++ packages/warriorjs-cli/src/layouts/components/totalScore.js
