+++ packages/warriorjs-cli/src/layouts/components/warriorStatus.js
